requirements.txt
pytest==8.4.2
pytest-playwright==0.7.1
playwright==1.56.0
python-dotenv==1.0.1
faker==33.0.0
